+++ src/compiler/transformers/component-lazy/lazy-constructor.ts
@@ -11,7 +11,7 @@ export const updateLazyComponentConstructor = (
   cmp: d.ComponentCompilerMeta
 ) => {
   const cstrMethodArgs = [
-    ts.createParameter(undefined, undefined, undefined, ts.factory.createIdentifier(HOST_REF_ARG)),
+    ts.factory.createParameterDeclaration(undefined, undefined, undefined, ts.factory.createIdentifier(HOST_REF_ARG)),
   ];
 
   const cstrMethodIndex = classMembers.findIndex((m) => m.kind === ts.SyntaxKind.Constructor);
@@ -19,14 +19,14 @@ export const updateLazyComponentConstructor = (
     // add to the existing constructor()
     const cstrMethod = classMembers[cstrMethodIndex] as ts.ConstructorDeclaration;
 
-    const body = ts.updateBlock(cstrMethod.body, [
+    const body = ts.factory.updateBlock(cstrMethod.body, [
       registerInstanceStatement(moduleFile),
       ...addCreateEvents(moduleFile, cmp),
       ...cstrMethod.body.statements,
       ...addLegacyProps(moduleFile, cmp),
     ]);
 
-    classMembers[cstrMethodIndex] = ts.updateConstructor(
+    classMembers[cstrMethodIndex] = ts.factory.updateConstructorDeclaration(
       cstrMethod,
       cstrMethod.decorators,
       cstrMethod.modifiers,
@@ -35,7 +35,7 @@ export const updateLazyComponentConstructor = (
     );
   } else {
     // create a constructor()
-    const cstrMethod = ts.createConstructor(
+    const cstrMethod = ts.factory.createConstructorDeclaration(
       undefined,
       undefined,
       cstrMethodArgs,
@@ -55,7 +55,7 @@ export const updateLazyComponentConstructor = (
 const registerInstanceStatement = (moduleFile: d.Module) => {
   addCoreRuntimeApi(moduleFile, RUNTIME_APIS.registerInstance);
 
-  return ts.createStatement(
+  return ts.factory.createExpressionStatement(
     ts.factory.createCallExpression(ts.factory.createIdentifier(REGISTER_INSTANCE), undefined, [
       ts.factory.createThis(),
       ts.factory.createIdentifier(HOST_REF_ARG),