+++ src/compiler/transformers/decorators-to-static/convert-decorators.ts
@@ -84,7 +84,7 @@ const visitClassDeclaration = (
 
   validateMethods(diagnostics, classMembers);
 
-  return ts.updateClassDeclaration(
+  return ts.factory.updateClassDeclaration(
     classNode,
     filterDecorators(classNode, CLASS_DECORATORS_TO_REMOVE),
     classNode.modifiers,
@@ -121,7 +121,7 @@ const removeStencilMethodDecorators = (
 
     if (currentDecorators !== newDecorators) {
       if (ts.isMethodDeclaration(member)) {
-        return ts.updateMethod(
+        return ts.factory.updateMethodDeclaration(
           member,
           newDecorators,
           member.modifiers,
@@ -143,7 +143,7 @@ const removeStencilMethodDecorators = (
           return member;
         } else {
           // update the property to remove decorators
-          return ts.updateProperty(
+          return ts.factory.updatePropertyDeclaration(
             member,
             newDecorators,
             member.modifiers,
@@ -372,10 +372,24 @@ export const updateConstructor = (
   const constructorMethod = classMembers[constructorIndex];
 
   if (constructorIndex >= 0 && ts.isConstructorDeclaration(constructorMethod)) {
-    const hasSuper = constructorMethod.body.statements.some((s) => s.kind === ts.SyntaxKind.SuperKeyword);
+    const constructorBodyStatements: ts.NodeArray<ts.Statement> =
+      constructorMethod.body?.statements ?? ts.factory.createNodeArray();
+    const hasSuper = constructorBodyStatements.some((s) => s.kind === ts.SyntaxKind.SuperKeyword);
 
     if (!hasSuper && needsSuper(classNode)) {
-      statements = [createConstructorBodyWithSuper(), ...statements];
+      // if there is no super and it needs one the statements comprising the
+      // body of the constructor should be:
+      //
+      // 1. the `super()` call
+      // 2. the new statements we've created to initialize fields
+      // 3. the statements currently comprising the body of the constructor
+      statements = [createConstructorBodyWithSuper(), ...statements, ...constructorBodyStatements];
+    } else {
+      // if no super is needed then the body of the constructor should be:
+      //
+      // 1. the new statements we've created to initialize fields
+      // 2. the statements currently comprising the body of the constructor
+      statements = [...statements, ...constructorBodyStatements];
     }
 
     classMembers[constructorIndex] = ts.factory.updateConstructorDeclaration(
@@ -383,7 +397,7 @@ export const updateConstructor = (
       constructorMethod.decorators,
       constructorMethod.modifiers,
       constructorMethod.parameters,
-      ts.factory.updateBlock(constructorMethod.body, statements)
+      ts.factory.updateBlock(constructorMethod?.body ?? ts.factory.createBlock([]), statements)
     );
   } else {
     // we don't seem to have a constructor, so let's create one and stick it
@@ -393,12 +407,7 @@ export const updateConstructor = (
     }
 
     classMembers = [
-      ts.factory.createConstructorDeclaration(
-        undefined,
-        undefined,
-        undefined,
-        ts.factory.createBlock(statements, true)
-      ),
+      ts.factory.createConstructorDeclaration(undefined, undefined, [], ts.factory.createBlock(statements, true)),
       ...classMembers,
     ];
   }