+++ src/compiler/transformers/remove-collection-imports.ts
@@ -42,7 +42,7 @@ export const removeCollectionImports = (compilerCtx: d.CompilerCtx): ts.Transfor
       }
 
       if (madeUpdates) {
-        return ts.updateSourceFileNode(tsSourceFile, statements);
+        return ts.factory.updateSourceFile(tsSourceFile, statements);
       }
       return tsSourceFile;
     };